Changed name of BinaryClock to BinaryClockPanel
